Implement scrobbling, improved media checks
Implemented scrobbling with the track.updateNowPlaying and track.scrobble APIs.

Improved the checks that a playing media is actually music. Most stuff coming from the browser is reported to Windows as music, and a lot of non-music youtube videos are in the LastFM database. However, most (if not all) music on LastFM has an album associated with it, and I've yet to find a non-music video with an album. So to verify that the "music" is actually music, we can check if LastFM associates it with an album.